requirements.txt
Flask==3.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
Flask-SocketIO==5.3.5
python-socketio==5.10.0
ultralytics==8.1.0
pygame==2.5.2
python-chess==1.999

# NEW DEPENDENCIES FOR IMPROVEMENTS
onnxruntime-gpu==1.16.3  # ONNX inference (faster than PyTorch)
onnx==1.15.0             # ONNX model format
filterpy==1.4.5          # Kalman filter for temporal smoothing
albumentations==1.3.1    # Advanced data augmentation (optional, for training only)
scipy==1.11.4            # Already used, but ensure version compatibility
scikit-learn==1.3.2      # Already used (DBSCAN clustering)

# OPTIONAL (for development/testing)
pytest==7.4.3            # Unit testing
tqdm==4.66.1             # Progress bars
